+++ src/inspectorcell/util/entitycli.py
@@ -127,7 +127,7 @@ def to_pixmap(args):
     
     def make_pixmap(xres, yres, tiffile):
         col_fn = lambda ent: ent.eid
-        img = np.zeros((xres, yres, 1), np.uint16)
+        img = np.zeros((yres, xres, 1), np.uint16)
         draw_entities(img, eman, col_fn, stroke=0)
         pic = Image.fromarray(img[..., 0])
         pic.save(tiffile)
@@ -135,6 +135,62 @@ def to_pixmap(args):
     tiffile = Path(args.outtif)
     wrapper = DryrunWrapper(args)
     wrapper.perform(make_pixmap, (), dict(xres=xres, yres=yres,
+                    tiffile=args.outtif))
+
+def draw_cluster(args):
+    eman = read_into_manager(args.injson, strip=True)
+    xres = int(args.xres)
+    yres = int(args.yres)
+    stroke = int(args.stroke)
+    
+    #TODO Make me a function and REUSE ME!
+    print('Clipping Entities...')
+    for ent in eman:
+        new_cont = []
+        for cont in ent.contours:
+            cur_cont = []
+            new_cont.append(cur_cont)
+            for (p0, p1) in cont:
+                p0 = min(max(p0, 0), xres)
+                p1 = min(max(p1, 0), yres)
+                cur_cont.append((p0, p1))
+
+        ent.from_contours(new_cont)
+    
+    def get_colorfun(colorcsv):
+        colscheme = pd.read_csv(colorcsv)
+        defaul = colscheme
+        LUT = {}
+        default = (50, 50, 50)
+        for i, r in colscheme.iterrows():
+            tag = r['tag']
+            rgb = r['red'], r['green'], r['blue']
+            if not tag == 'default':
+                LUT[tag] = rgb 
+            if tag == 'default':
+                default = rgb
+
+        def _colfn(ent):
+            ret = default
+            for curt in ent.tags:
+                ret = LUT.get(curt, default)
+                if ret != default:
+                    break
+            return ret
+
+        return _colfn
+
+    def make_pixmap(xres, yres, stroke, colorcsv, tiffile):
+        col_fn = get_colorfun(colorcsv)
+        img = np.zeros((yres, xres, 3), np.uint8)
+        draw_entities(img, eman, col_fn, stroke=stroke)
+        pic = Image.fromarray(img)
+        pic.save(tiffile)
+    
+    tiffile = Path(args.outtif)
+    wrapper = DryrunWrapper(args)
+    wrapper.perform(make_pixmap, (), dict(xres=xres, yres=yres,
+                    stroke=stroke, colorcsv=args.colorcsv,
                     tiffile=args.outtif))
 
 
@@ -196,6 +252,28 @@ def main(*args, **kwargs):
     totif.add_argument('-n', '--dryrun', action='store_true')
     totif.set_defaults(func=to_pixmap)
 
+    # Make pixmap with color based on clustering
+    drawme = command_parser.add_parser(
+        'clusterdraw',
+         description='Draws Segments based on the clustering',
+        )
+    
+    drawme.add_argument('injson', type=str, help='json file to read Entities from')
+    drawme.add_argument('outtif', type=str, help='tif file to write to')
+    drawme.add_argument('colorcsv', type=str, help='CSV with coloring information')
+    drawme.add_argument('--stroke', type=int, help='Stroke width', nargs='?',
+                        default=0)
+    drawme.add_argument('--xres', type=int, help='tif pixel width', nargs='?',
+                        default=2048)
+    drawme.add_argument('--yres', type=int, help='tif pixel hight', nargs='?',
+                       default=2048)
+
+    drawme.add_argument('-n', '--dryrun', action='store_true')
+    drawme.set_defaults(func=draw_cluster)
+
+
+    
+    # Done
     parsed_args = entitycli.parse_args(sys.argv[1:])
     if hasattr(parsed_args, 'func'):
         parsed_args.func(parsed_args)